Install solidus without javascript compilers (webpacker)
Using --skip-webpack-install is only preventing the Rails install generator from 
running webacker:install, but the webpacker gem is still added to the Gemfile.

In the context of a fresh app that uses the bare minimum set of tools to run Solidus
we don't need that. People can still add it later if needed.