Updated Deploy to Heroku
Included warning section on database selection. If a user attempted to
install their application on Heroku utilizing rail's default database
sqlite3, they would encounter an unexpected error. The warning is a
necessary brief notification to ensure the user is aware of this prior
to attempting to upload their application to Heroku.