Update Testing Solidus guide
It now takes into account that installing Solidus with the
Starter Frontend, users will have a test suite ready in their
application, which means that setting up the test environment
from scratch is not always requried.

This change also align the guide with what's provided in the
starter frontend, so that user that still want to read the
whole page to learn more, as suggested, will find similar things
in the guides and in their application.